# openapi.yaml
openapi: 3.0.3
info:
  title: Professional Pride Room Booking API
  version: 2.0.0
  description: |
    Comprehensive REST API for room booking management system with advanced features including:
    - Room availability validation and conflict prevention
    - Customer management with ID proof verification
    - Dynamic pricing based on room configurations
    - Excel-style booking grid with horizontal scrolling
    - Payment processing with screenshot uploads
    - Invoice generation and PDF export
    - Real-time dashboard analytics
    - Mobile-optimized responsive interfaces
    
    **Key Features:**
    - Prevents double bookings with real-time validation
    - Supports multiple ID proof uploads per customer
    - Dynamic pricing based on number of guests
    - Comprehensive booking history tracking
    - Professional invoice generation
    - Mobile-responsive design
servers:
  - url: https://your-service.onrender.com
    description: Production server
  - url: http://localhost:8082
    description: Local development server
paths:
...
  /api/v1/payments:
    get:
      summary: List payments
      responses:
        '200': 
          description: List of payments
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Payment'
    post:
      summary: Create payment
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201': 
          description: Payment created successfully
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/Payment'
  /api/v1/payments/{id}:
    get:
      summary: Get payment
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '200': 
          description: Payment details
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/Payment'
        '404': 
          description: Payment not found
    put:
      summary: Update payment
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/PaymentUpdateRequest'
      responses:
        '200': 
          description: Payment updated successfully
    delete:
      summary: Delete payment
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '204': 
          description: Payment deleted successfully
  /api/v1/payments/booking/{bookingId}:
    get:
      summary: Get payments for booking
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: 
            type: integer
            format: int64
      responses:
        '200': 
          description: Booking payments
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Payment'
...
  /api/v1/dashboard/today-summary:
    get:
      summary: Get today's summary
      description: Get comprehensive dashboard summary for a specific date
      parameters:
        - in: query
          name: date
          required: true
          schema: 
            type: string
            format: date
          description: Date for summary (YYYY-MM-DD)
      responses:
        '200': 
          description: Today's summary
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/DashboardSummary'
  /api/v1/dashboard/recent-bookings:
    get:
      summary: Get recent bookings
      parameters:
        - in: query
          name: limit
          required: false
          schema: 
            type: integer
            default: 10
          description: Number of recent bookings to return
      responses:
        '200': 
          description: Recent bookings
          content: 
            application/json: 
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Booking'
...
components:
  schemas:
    Booking:
      type: object
      properties:
        id: 
          type: integer
          format: int64
          description: Unique booking identifier
        customerPhoneNumber: 
          type: string
          description: Customer phone number
        roomId: 
          type: integer
          format: int64
          description: Room identifier
        numberOfPeople: 
          type: integer
          description: Number of guests
        checkInDate: 
          type: string
          format: date-time
          description: Check-in date and time
        checkOutDate: 
          type: string
          format: date-time
          description: Check-out date and time
        bookingDurationType: 
          $ref: '#/components/schemas/BookingDurationType'
        bookingStatus: 
          $ref: '#/components/schemas/BookingStatus'
        paymentStatus: 
          $ref: '#/components/schemas/PaymentStatus'
        dailyCost: 
          type: number
          format: double
          description: Daily rate cost
        monthlyCost: 
          type: number
          format: double
          description: Monthly rate cost
        earlyCheckinCost: 
          type: number
          format: double
          description: Early check-in additional cost
        lateCheckoutCost: 
          type: number
          format: double
          description: Late check-out additional cost
        totalAmount: 
          type: number
          format: double
          description: Total booking amount
        remarks: 
          type: string
          description: Additional remarks
        createdAt: 
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt: 
          type: string
          format: date-time
          description: Last update timestamp
...
    Payment:
      type: object
      properties:
        id: 
          type: integer
          format: int64
        bookingId: 
          type: integer
          format: int64
        amount: 
          type: number
          format: double
          description: Payment amount
        paymentMethod: 
          $ref: '#/components/schemas/PaymentMode'
        paymentStatus: 
          $ref: '#/components/schemas/PaymentStatus'
        paymentScreenshotUrl: 
          type: string
          description: Screenshot of payment proof
        transactionId: 
          type: string
          description: Transaction reference ID
        paymentDate: 
          type: string
          format: date-time
          description: Payment date and time
        createdAt: 
          type: string
          format: date-time
        updatedAt: 
          type: string
          format: date-time
    PaymentCreateRequest:
      type: object
      required: [bookingId, amount, paymentMethod]
      properties:
        bookingId: 
          type: integer
          format: int64
        amount: 
          type: number
          format: double
        paymentMethod: 
          $ref: '#/components/schemas/PaymentMode'
        paymentStatus: 
          $ref: '#/components/schemas/PaymentStatus'
        paymentScreenshotUrl: 
          type: string
        transactionId: 
          type: string
        paymentDate: 
          type: string
          format: date-time
    PaymentUpdateRequest:
      type: object
      properties:
        amount: 
          type: number
          format: double
        paymentMethod: 
          $ref: '#/components/schemas/PaymentMode'
        paymentStatus: 
          $ref: '#/components/schemas/PaymentStatus'
        paymentScreenshotUrl: 
          type: string
        transactionId: 
          type: string
        paymentDate: 
          type: string
          format: date-time
...
    DashboardSummary:
      type: object
      properties:
        date: 
          type: string
          format: date
        totalBookings: 
          type: integer
        confirmedBookings: 
          type: integer
        checkedInBookings: 
          type: integer
        checkedOutBookings: 
          type: integer
        totalRevenue: 
          type: number
          format: double
        checkIns: 
          type: array
          items: 
            $ref: '#/components/schemas/Booking'
          description: Today's check-ins
        checkOuts: 
          type: array
          items: 
            $ref: '#/components/schemas/Booking'
          description: Today's check-outs
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, CHECKEDIN, CHECKEDOUT, CANCELLED, NO_SHOW, COMPLETED]
      description: |
        PENDING - Booking request submitted
        CONFIRMED - Booking confirmed by staff
        CHECKEDIN - Customer has checked in
        CHECKEDOUT - Customer has checked out
        CANCELLED - Booking cancelled
        NO_SHOW - Customer did not show up
        COMPLETED - Booking completed successfully
    BookingDurationType:
      type: string
      enum: [DAILY, MONTHLY]
      description: |
        DAILY - Daily rate booking
        MONTHLY - Monthly rate booking
    PaymentStatus:
      type: string
      enum: [PENDING, PAID, FAILED]
      description: |
        PENDING - Payment pending
        PAID - Payment completed
        FAILED - Payment failed
    PaymentMode:
      type: string
      enum: [CASH, UPI, CARD, BANK_TRANSFER]
      description: |
        CASH - Cash payment
        UPI - UPI payment
        CARD - Card payment
        BANK_TRANSFER - Bank transfer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - BearerAuth: []